+++ src/containers/SidePanelDetails/DataAccessReports/index.js
@@ -4,7 +4,12 @@
 
 import { connect } from 'react-redux';
 import DataAccessReportsDetailsPanel from './layout';
-import { userHasUpdatePermission, getCurrentEntity, itemApiPending } from '../../../redux/modules/entities/selectors';
+import {
+  userHasUpdatePermission,
+  getCurrentEntity,
+  isEntityFetching,
+  itemApiPending
+} from '../../../redux/modules/entities/selectors';
 
 import { selectEntityListMembers, filterUsersByPermissions } from '../../../redux/modules/entities/users/selectors';
 import { setSelectedSubEntityId, removeListItem } from '../../../redux/modules/entities';
@@ -17,6 +22,7 @@ export function mapStateToProps(state, props) {
     userHasUpdatePermission: userHasUpdatePermission(state),
     tableItems: filterUsersByPermissions(state, listMembers, ['ASSIGNED_REPORTS_READ']),
     tableFields: entitiesMetaData[currentEntity].sidePanelListTableFields,
+    usersFetching: isEntityFetching(state, 'users'),
     itemApiPending: itemApiPending(state),
     defaultFilters: entitiesMetaData[currentEntity].defaultDependentEntityFilters
   };